comment: move ethflow into swap
