--- src/cow-react/modules/ethFlow/index.tsx
@@ -1,1 +1,0 @@
-export { EthFlowModal, type EthFlowProps } from './containers/EthFlowModal'
