--- src/cow-react/modules/ethFlow/pure/WrappingPreview/index.tsx
@@ -1,1 +1,0 @@
-export { WrappingPreview, type WrappingPreviewProps } from './WrappingPreview'
