+++ src/cow-react/modules/swap/containers/EthFlow/hooks/useEthFlowActions.ts
@@ -2,7 +2,7 @@ import { useMemo } from 'react'
 import { Field } from 'state/swap/actions'
 import { WRAPPED_NATIVE_CURRENCY } from 'constants/tokens'
 import { useAtomValue } from 'jotai/utils'
-import { ethFlowContextAtom, updateEthFlowContextAtom } from '../../../state/ethFlowContextAtom'
+import { ethFlowContextAtom, updateEthFlowContextAtom } from '../../../state/EthFlow/ethFlowContextAtom'
 import { useSetAtom } from 'jotai'
 import { useDerivedSwapInfo, useSwapActionHandlers } from 'state/swap/hooks'
 import { useSwapConfirmManager } from 'cow-react/modules/swap/hooks/useSwapConfirmManager'
