+++ src/cow-react/modules/swap/containers/EthFlow/index.tsx
@@ -8,15 +8,15 @@ import { ApprovalState, ApproveCallback } from 'hooks/useApproveCallback'
 import { WrapUnwrapCallback } from 'hooks/useWrapCallback'
 import { useDetectNativeToken } from 'state/swap/hooks'
 import { GpModal } from 'components/Modal'
-import { EthFlowModalContent } from '../../pure/EthFlowModalContent'
-import { getDerivedEthFlowState } from './utils/getDerivedEthFlowState'
-import { ethFlowContextAtom } from '../../state/ethFlowContextAtom'
+import { EthFlowModalContent } from '@cow/modules/swap/pure/EthFlow/EthFlowModalContent'
+import { getDerivedEthFlowState } from '@cow/modules/swap/containers/EthFlow/utils/getDerivedEthFlowState'
+import { ethFlowContextAtom } from '@cow/modules/swap/state/EthFlow/ethFlowContextAtom'
 import { useAtomValue } from 'jotai/utils'
-import { WrappingPreviewProps } from '../../pure/WrappingPreview'
+import { WrappingPreviewProps } from '@cow/modules/swap/pure/EthFlow/WrappingPreview'
 import { useSingleActivityDescriptor } from 'hooks/useRecentActivity'
 import { useEthFlowActions } from './hooks/useEthFlowActions'
 import { useSetupEthFlow } from './hooks/useSetupEthFlow'
-import { HandleSwapCallback } from 'cow-react/modules/swap/hooks/useHandleSwap'
+import { HandleSwapCallback } from '@cow/modules/swap/hooks/useHandleSwap'
 
 export interface EthFlowProps {
   nativeInput?: CurrencyAmount<Currency>
