+++ src/cow-react/modules/swap/pure/EthFlow/WrappingPreview/WrappingPreview.cosmos.tsx
@@ -1,6 +1,5 @@
-import { WrappingPreview } from '.'
-
-import { wrappingPreviewProps } from '../../mocks'
+import { WrappingPreview } from '@cow/modules/swap/pure/EthFlow/WrappingPreview'
+import { wrappingPreviewProps } from '@cow/modules/swap/services/ethFlow/mocks'
 
 const Fixtures = {
   withBalance: <WrappingPreview {...wrappingPreviewProps} />,
