+++ src/cow-react/modules/swap/pure/EthFlow/WrappingPreview/index.tsx
@@ -1,0 +1,4 @@
+export {
+  WrappingPreview,
+  type WrappingPreviewProps,
+} from '@cow/modules/swap/pure/EthFlow/WrappingPreview/WrappingPreview'
