+++ src/cow-react/pages/Swap/SwapMod.tsx
@@ -55,7 +55,7 @@ import { ArrowWrapperLoader } from 'components/ArrowWrapperLoader'
 import { HighFeeWarning, NoImpactWarning } from 'components/SwapWarnings'
 import { FeesDiscount } from '@cow/modules/swap/containers/FeesDiscount'
 import { RouteComponentProps } from 'react-router-dom'
-import { EthFlowModal } from '@cow/modules/ethFlow'
+import { EthFlowModal } from '@cow/modules/swap/containers/EthFlow'
 import { useSwapButtonContext } from '@cow/modules/swap/hooks/useSwapButtonContext'
 import { Routes } from '@cow/constants/routes'
 import { PageTitle } from '@cow/modules/application/containers/PageTitle'
